+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 2.x branch
 ## 2.1 branch
+### 2.1.4
+* improved related posts. Now you can also show images;
+* added the User shell, to manage users;
+* added backward compatibility for old URLs;
+* fixed a serious bug for static pages.
+
 ### 2.1.3
 * added related posts for each post;
 * now you can list and edit tags and lists posts by tags;